+++ enum/src/main/java/org/grnet/status/enums/EventName.java
@@ -2,5 +2,6 @@ package org.grnet.status.enums;
 
 public enum EventName {
         INIT_AMS,
-        INIT_MONGO
+        INIT_MONGO,
+        INIT_COMPUTE_ENGINE
 }